requirements.txt
flask
langchain
langchain-community
langchain-core
chromadb
pypdf
langchain-groq
python-dotenv
ipykernel
sentence-transformers
